+++ CHANGELOG.md
@@ -8,6 +8,8 @@ All dates are in the European Central timezone.
 Version 12.2 *(in development)*
 ---------------------------
 
+* Comments for episodes did wrongly display show comments.
+
 #### 12.2-beta1 *(2014-03-01)*
 
 * FIX Crashes when doing (bulk) database operations. This might have performance impacts regarding adding shows, sync time, etc. Please report anything out of the ordinary in that regard.